Refactored filter generation
Also added an ENABLE_BOT preprocesser directive. If this is removed, the bot functionality will not be compiled in, which is useful for testing.

Additionally, merophony/holophony and merography/holography were mistakenly backwards in verbly before, and this has been updated. This should help with some of the "query returned empty dataset" cases.

refs #7